Add NodeCallback.
